+++ tilings/strategies/factor.py
@@ -215,27 +215,6 @@ class Interleaving(CartesianProduct[Tiling, GriddedPerm]):
         super().__init__(parent, children, extra_parameters)
         self.interleaving_parameters = tuple(interleaving_parameters)
 
-    @classmethod
-    def untracked(
-        cls,
-        parent: Tiling,
-        children: Iterable[Tiling],
-        extra_parameters: Tuple[Dict[str, str], ...],
-    ) -> "Interleaving":
-        res = cls(parent, children, extra_parameters, [])
-
-        def f(*args, **kwargs):
-            raise NotImplementedError(
-                "enumeration for untracked interleaving factors is not implemented"
-            )
-
-        # Overwriting methods is bad practice, but useful here!
-        res.get_recurrence = f  # type: ignore
-        res.get_equation = f  # type: ignore
-        res.get_sub_objects = f  # type: ignore
-        res.random_sample_sub_objects = f  # type: ignore
-        return res
-
     @staticmethod
     def is_equivalence() -> bool:
         return False
@@ -267,17 +246,9 @@ class Interleaving(CartesianProduct[Tiling, GriddedPerm]):
         subsamplers: SubSamplers,
         subrecs: SubRecs,
         n: int,
-        **parameters: int
+        **parameters: int,
     ):
         raise NotImplementedError
-
-    @staticmethod
-    def get_eq_symbol() -> str:
-        return "="
-
-    @staticmethod
-    def get_op_symbol() -> str:
-        return "*"
 
 
 class FactorWithInterleavingStrategy(FactorStrategy):
@@ -293,9 +264,7 @@ class FactorWithInterleavingStrategy(FactorStrategy):
             interleaving_parameters = self.interleaving_parameters(tiling)
         except ValueError:
             # must be untracked
-            return Interleaving.untracked(
-                tiling, children, self.extra_parameters(tiling, children)
-            )
+            raise NotImplementedError("The interleaving factor was not tracked.")
         return Interleaving(
             tiling,
             children,
@@ -347,6 +316,14 @@ class FactorWithInterleavingStrategy(FactorStrategy):
     ) -> Tuple[GriddedPerm, ...]:
         raise NotImplementedError
 
+    @staticmethod
+    def get_eq_symbol() -> str:
+        return "="
+
+    @staticmethod
+    def get_op_symbol() -> str:
+        return "*"
+
     def __repr__(self) -> str:
         return (
             self.__class__.__name__
@@ -368,12 +345,7 @@ class FactorWithMonotoneInterleavingStrategy(FactorWithInterleavingStrategy):
             interleaving_parameters = self.interleaving_parameters(tiling)
         except ValueError:
             # must be untracked
-            return cast(
-                MonotoneInterleaving,
-                MonotoneInterleaving.untracked(
-                    tiling, children, self.extra_parameters(tiling, children)
-                ),
-            )
+            raise NotImplementedError("The monotone interleaving was not tracked.")
         return MonotoneInterleaving(
             tiling,
             children,
@@ -431,23 +403,35 @@ class FactorFactory(StrategyFactory[Tiling]):
                     if not self.tracked or contains_interleaving_assumptions(
                         comb_class, components
                     ):
-                        yield self.factor_class(
-                            components, ignore_parent=self.ignore_parent, workable=False
-                        )
+                        yield self._build_strategy(components, workable=False)
             if not self.tracked or contains_interleaving_assumptions(
                 comb_class, min_comp
             ):
-                yield self.factor_class(
-                    min_comp, ignore_parent=self.ignore_parent, workable=self.workable,
+                yield self._build_strategy(
+                    min_comp, workable=self.workable,
                 )
+
+    def _build_strategy(
+        self, components: Tuple[Tuple[Cell, ...], ...], workable: bool
+    ) -> FactorStrategy:
+        """
+        Build the factor strategy for the given components.
+
+        It ensure that a plain factor rule is returned.
+        """
+        interleaving = any(interleaving_rows_and_cols(components))
+        factor_strat = self.factor_class if interleaving else FactorStrategy
+        return factor_strat(
+            components, ignore_parent=self.ignore_parent, workable=workable,
+        )
 
     def __str__(self) -> str:
         if self.factor_class is FactorStrategy:
             s = "factor"
         elif self.factor_class is FactorWithInterleavingStrategy:
-            s = "factor with monotone interleaving"
-        elif self.factor_class is FactorWithMonotoneInterleavingStrategy:
             s = "factor with interleaving"
+        elif self.factor_class is FactorWithMonotoneInterleavingStrategy:
+            s = "factor with monotone interleaving"
         else:
             raise Exception("Invalid interleaving type")
         if self.unions: